+++ packages/rfw/test/argument_decoders_test.dart
@@ -352,6 +352,7 @@ void main() {
     await expectLater(
       find.byType(RemoteWidget),
       matchesGoldenFile('goldens/argument_decoders_test.containers.png'),
+      skip: 'https://github.com/flutter/flutter/issues/106205'
     );
     expect(find.byType(DecoratedBox), findsNWidgets(6));
     expect(
@@ -427,6 +428,7 @@ void main() {
     await expectLater(
       find.byType(RemoteWidget),
       matchesGoldenFile('goldens/argument_decoders_test.text.png'),
+      skip: 'https://github.com/flutter/flutter/issues/106205'
     );
 
     runtime.update(const LibraryName(<String>['test']), parseLibraryFile('''
@@ -449,6 +451,7 @@ void main() {
     await expectLater(
       find.byType(RemoteWidget),
       matchesGoldenFile('goldens/argument_decoders_test.gridview.fixed.png'),
+      skip: 'https://github.com/flutter/flutter/issues/106205'
     );
 
     runtime.update(const LibraryName(<String>['test']), parseLibraryFile('''
@@ -471,6 +474,7 @@ void main() {
     await expectLater(
       find.byType(RemoteWidget),
       matchesGoldenFile('goldens/argument_decoders_test.gridview.max.png'),
+      skip: 'https://github.com/flutter/flutter/issues/106205'
     );
 
     int sawGridDelegateDecoder = 0;
@@ -500,6 +504,7 @@ void main() {
     await expectLater(
       find.byType(RemoteWidget),
       matchesGoldenFile('goldens/argument_decoders_test.gridview.custom.png'),
+      skip: 'https://github.com/flutter/flutter/issues/106205'
     );
   }, skip: !runGoldens);
 }