+++ packages/pigeon/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 3.1.7
+
+* [java] Adds option to add javax.annotation.Generated annotation.
+
+## 3.1.6
+
+* Supports newer versions of `analyzer`.
+
 ## 3.1.5
 
 * Fixes potential crash bug when using a nullable nested type that has nonnull